feature(feedback): Update Unknown Component and Refactor Username callback
Add the Unknown Component that I missed completely, and Update the
Username component to be the same as the Text field component
